Move Mongo access and structs into separate packages
