+++ handlers/public.go
@@ -16,53 +16,6 @@ import (
 	"github.com/gorilla/pat"
 )
 
-type incomingPaymentResourceRequest struct {
-	RedirectURI string `json:"redirect_uri"`
-	Reference   string `json:"reference"`
-	Resource    string `json:"resource"`
-	State       string `json:"state"`
-}
-
-// PaymentResource is the payment details to be stored in the DB and returned in the response
-type PaymentResource struct {
-	Amount                  string    `json:"amount"                              bson:"amount"`
-	AvailablePaymentMethods []string  `json:"available_payment_methods,omitempty" bson:"available_payment_methods,omitempty"`
-	CompletedAt             time.Time `json:"completed_at,omitempty"              bson:"completed_at,omitempty"`
-	CreatedAt               time.Time `json:"created_at,omitempty"                bson:"created_at,omitempty"`
-	CreatedBy               CreatedBy `json:"created_by"                          bson:"created_by"`
-	Description             string    `json:"description"                         bson:"description"`
-	Links                   Links     `json:"links"                               bson:"links"`
-	PaymentMethod           string    `json:"payment_method,omitempty"            bson:"payment_method,omitempty"`
-	Reference               string    `json:"reference,omitempty"                 bson:"reference,omitempty"`
-	Status                  string    `json:"status"                              bson:"status"`
-}
-
-// CreatedBy is the user who is creating the payment session
-type CreatedBy struct {
-	Email    string `bson:"email"`
-	Forename string `bson:"forename"`
-	ID       string `bson:"id"`
-	Surname  string `bson:"surname"`
-}
-
-// Links is a set of URLs related to the resource, including self
-type Links struct {
-	Journey  string `json:"journey"`
-	Resource string `json:"resource"`
-	Self     string `json:"self"`
-}
-
-// Data is a representation of the top level data retrieved from the Transaction API
-type Data struct {
-	CompanyName string            `json:"company_name"`
-	Filings     map[string]Filing `json:"filings"`
-}
-
-// Filing is a representation of the Filing subsection of data retrieved from the Transaction API
-type Filing struct {
-	Description string `json:"description"`
-}
-
 // Register defines the route mappings
 func Register(r *pat.Router) {
 	r.Get("/healthcheck", getHealthCheck).Name("get-healthcheck")
@@ -83,7 +36,7 @@ func createPaymentSession(w http.ResponseWriter, req *http.Request) {
 	}
 
 	requestDecoder := json.NewDecoder(req.Body)
-	var incomingPaymentResourceRequest incomingPaymentResourceRequest
+	var incomingPaymentResourceRequest data.IncomingPaymentResourceRequest
 	err := requestDecoder.Decode(&incomingPaymentResourceRequest)
 	if err != nil {
 		log.ErrorR(req, fmt.Errorf("Request Body Invalid"))
@@ -142,7 +95,7 @@ func createPaymentSession(w http.ResponseWriter, req *http.Request) {
 
 	// TODO save cost resource and ensure all mandatory fields are present:
 
-	paymentResource := &PaymentResource{}
+	paymentResource := &data.PaymentResource{}
 	err = json.Unmarshal(body, paymentResource)
 	if err != nil {
 		log.ErrorR(resourceReq, fmt.Errorf("Error reading Cost Resource: [%s]", err))
@@ -167,7 +120,7 @@ func createPaymentSession(w http.ResponseWriter, req *http.Request) {
 			return
 		}
 	}
-	paymentResource.CreatedBy = CreatedBy{
+	paymentResource.CreatedBy = data.CreatedBy{
 		ID:       req.Header.Get("Eric-Identity"),
 		Email:    email,
 		Forename: forename,
@@ -177,20 +130,9 @@ func createPaymentSession(w http.ResponseWriter, req *http.Request) {
 	paymentResource.CreatedAt = time.Now()
 	paymentResource.Reference = incomingPaymentResourceRequest.Reference
 
-	// Write to DB
-	session, err := data.GetMongoSession()
-	if err != nil {
-		log.ErrorR(req, fmt.Errorf("Error connecting to MongoDB: %s", err))
-		w.WriteHeader(http.StatusInternalServerError) // 500
-		return
-	}
-	defer session.Close()
-
-	c := session.DB("transactions").C("payments")
-
-	if err = c.Insert(paymentResource); err != nil {
-		log.ErrorR(req, fmt.Errorf("Error writing to MongoDB: %s", err))
-		w.WriteHeader(http.StatusInternalServerError) // 500
+	if data.CreatePaymentResourceDB(paymentResource) != nil {
+		log.ErrorR(req, fmt.Errorf("error writing to MongoDB: %s", err))
+		w.WriteHeader(http.StatusInternalServerError)
 		return
 	}
